# gym-app/src/main/java/com/epam/gym/main/dto/TraineeResponse.java
package com.epam.gym.main.dto;

import lombok.Builder;

import java.util.List;

@Builder
public record TraineeResponse(
        String username,
        String firstName,
        String lastName,
        String dateOfBirth,
        String address,
        boolean isActive,
        List<BasicTrainerResponse> trainers
) {
}